Scroll timestamp into view only if it's far from the center.
This avoids constant scrolling when the viewport is large enough to keep a significant window.

But when we do scroll, we scroll into the center. This allows subsequent scrolls to be skipped more often.

The main consequence is that the scrolling is less likely to interrupt someone actively in the process of trying to select a timestamp. A uture improvement would be to avoid scrolling if the user has scrolled in the last X seconds (and an auto-scroll has not happened since).